requirements.txt
# Backend dependencies for Qodo Merge testing project
# Some versions are intentionally outdated for testing purposes

fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.0
SQLAlchemy==1.4.0
psycopg2-binary==2.9.5
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.2.0
pytest-asyncio==0.20.3
requests==2.28.0
python-dotenv==1.0.0

# INTENTIONAL ISSUE: Outdated packages with known vulnerabilities
Pillow==8.1.1  # Has known security vulnerabilities
PyYAML==5.4.1  # Outdated version
urllib3==1.26.20  # Outdated version

# INTENTIONAL ISSUE: Packages not pinned to specific versions
numpy
pandas
matplotlib